# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QR Register SaaS API
  version: 1.0.0
  description: Contrat API versionné pour le registre visiteurs et ses espaces d’équipe.
paths:
...
  /api/v1/admin/organizations/:
    get:
      operationId: v1_admin_organizations_retrieve
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    post:
      operationId: v1_admin_organizations_create
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/admin/organizations/{organization_id}/:
    patch:
      operationId: v1_admin_organizations_partial_update
      parameters:
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    delete:
      operationId: v1_admin_organizations_destroy
      parameters:
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT